# ArtoriasCore/src/Swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
      "description": "This is a simple example NodeJS API project to demonstrate Swagger Documentation",
      "version": "1.0.0",
      "title": "Users API",
      "contact": {
          "email": "abc@gmail.com"
      },
      "license": {
          "name": "Apache 2.0",
          "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
  },
  "schemes": ["http"],
  "host": "localhost:3081",
  "basePath": "/api",
  "paths" : {
      "/users" : {
          "get" : {
              "summary" : "Get all the users",
              "description": "Get all the users",
              "produces": ["application/json"],
              "parameters": [],
              "responses": {
                  "200": {
                      "description": "successful operation",
                      "schema": {
                          "type": "array",
                          "items": {
                              "$ref": "#/definitions/usersResponse"
                          }
                      }
                  },
                  "400": {
                      "description": "Invalid status value",
                      "schema": {
                          "$ref": "#/definitions/InvalidResponse"
                      }
                  }
              }
          }
      },
      "/user" : {
          "post" : {
              "summary" : "Save the user",
              "description": "Save the user",
              "produces": ["application/json"],
              "consumes": ["application/json"],
              "parameters": [
                  {
                      "in": "body",
                      "name": "body",
                      "description": "user object",
                      "required": true,
                      "schema": {
                          "type": "object",
                          "properties": {
                              "user" : {
                                  "type": "object",
                                  "$ref": "#/definitions/User"
                              } 
                          }
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "successful operation",
                      "schema": {
                          "type": "array",
                          "items": {
                              "$ref": "#/definitions/usersResponse"
                          }
                      }
                  },
                  "400": {
                      "description": "Invalid status value",
                      "schema": {
                          "$ref": "#/definitions/InvalidResponse"
                      }
                  }
              }
          }
      },
      "/user/" : {
          "put" : {
              "summary" : "Update the users",
              "description": "Update the users",
              "produces": ["application/json"],
              "parameters": [
                  {
                      "name": "id",
                      "in": "path",
                      "description": "user id that needs to be updated",
                      "required": true,
                      "type": "string"
                  },
                  {
                      "in": "body",
                      "name": "body",
                      "description": "user object",
                      "required": true,
                      "schema": {
                          "type": "object",
                          "properties": {
                              "user" : {
                                  "type": "object",
                                  "$ref": "#/definitions/User"
                              } 
                          }
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "successful operation",
                      "schema": {
                          "type": "array",
                          "items": {
                              "$ref": "#/definitions/usersResponse"
                          }
                      }
                  },
                  "400": {
                      "description": "Invalid status value",
                      "schema": {
                          "$ref": "#/definitions/InvalidResponse"
                      }
                  }
              }
          }
      },
      "/user/{id}" : {
          "delete" : {
              "summary" : "Delete the user",
              "description": "Delete the user",
              "produces": ["application/json"],
              "parameters": [
                  {
                      "name": "id",
                      "in": "path",
                      "description": "user id that needs to be deleted",
                      "required": true,
                      "type": "string"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "successful operation",
                      "schema": {
                          "type": "array",
                          "items": {
                              "$ref": "#/definitions/usersResponse"
                          }
                      }
                  },
                  "400": {
                      "description": "Invalid status value",
                      "schema": {
                          "$ref": "#/definitions/InvalidResponse"
                      }
                  }
              }
          }
      }
  }, 
  "definitions": {
      "usersResponse": {
          "type": "object",
          "properties": {
              "_id": {
                   "type": "integer"
              },
              "first_name": {
                  "type": "string"
              },
              "last_name": {
                  "type": "string"
              },
              "email": {
                "type": "string"
              },
              "password": {
                "type": "string"
              },
              "createdBy": {
                  "type": "string"
              },
              "createdAt": {
                  "type": "string"
              },
              "updatedAt": {
                  "type": "string"
              },
              "updatedBy": {
                  "type": "string"
              }
          }
      },
      "User": {
          "type": "object",
          "properties": {
              "user": {
                  "type": {
                      "first_name": {
                          "type": "string"
                      },
                      "last_name": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "password": {
                        "type": "string"
                      },
                      "createdBy": {
                          "type": "string"
                      }
                  }
              }
          }
      },
      "InvalidResponse": {
          "type": "object",
          "properties": {
              "statusCode": {
                  "type": "string"
              },
              "message": {
                  "type": "string"
              }
          }

      }
  }
}